Use `WorkDiffTime` for `draftDuration`
With this, the caller of `draftDuration` can now trivially get the
"work" draft duration (excluding weekends).

The return type of `draftDuration` now wrapped in `Maybe` makes the
signature more correct: the function returns `Nothing` where there is no
draft time instead of 0.

I'm not sure if there is a simpler way to amend the tests than
introducing `draftDuration'` so that there is no need to call
`WorkTime.regular` in each test.