Download team's PRs from multiple repos
This removes downloading PRs by label, replacing it with downloading PRs
by author from multiple repositories.

Note: I removed the `instance Ord Pull` because the key is now "repo +
number" as it's not clear that the ascending ordering of `(repo,
number)` is always applicable. For example, `sortByRepoAndNumberDesc`
needs to order `Pull`s by ascending `repo` and descending `number`; and
the newest PR is now found by an explicit maximum created date instead
of relying on `Ord`.